requirements.txt
Django==3.0.6
asgiref==3.2.7
dj-database-url==0.5.0
# django-heroku==0.3.1
django-common-helpers==0.9.2
django-cron==0.5.1
# django-configurations==2.2
djangorestframework==3.11.0
djoser==2.0.3
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
